drivers: flash: expose spi_nor helpers in private API
To accomodate devices that extend the generic SPI NOR interface, some
functionality is convenient to expose. We place the resulting header in
the source directory, rather than in include/ since this is a
driver-facing API.

Examples include IC:s that expose special OTP (One Time Programmable)
registers. The implementational details of this feature does not seem
standardized, which means placing it in the generic spi_nor driver seems
like a bad idea.

Signed-off-by: Benjamin Lindqvist <benjamin.lindqvist@endian.se>